Move production code out.
